requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-js-asset==1.2.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
kavenegar==1.1.2
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
mysqlclient==2.0.1